+++ src/df-architect.ts
@@ -23,6 +23,7 @@ import LightCounter from './lights/LightCounter.js';
 import WallsCounter from './walls/WallCounter.js';
 import TileCounter from './tiles/TileCounter.js';
 import SoundCounter from './sounds/SoundCounter.js';
+import ShowLayerControls from './general/ShowLayerControls.js';
 
 Hooks.once('init', function () {
 	if (!game.modules.get('lib-wrapper')?.active) return;
@@ -41,6 +42,8 @@ Hooks.once('init', function () {
 
 	ARCHITECT.DrawArchitectGraphicToConsole();
 	try { DataMigration.init() } catch (error) { console.error(error) }
+	// WARNING! ShowLayerControls needs to come before AltGridSnap
+	try { ShowLayerControls.init() } catch (error) { console.error(error) }
 	try { AltGridSnap.init() } catch (error) { console.error(error) }
 	try { WallCtrlInvert.init() } catch (error) { console.error(error) }
 	try { WallJoinSplit.init() } catch (error) { console.error(error) }
@@ -50,10 +53,6 @@ Hooks.once('init', function () {
 	try { CaptureGameScreen.init() } catch (error) { console.error(error) }
 	try { LightTemplateManager.init() } catch (error) { console.error(error) }
 	try { TileFlattener.init() } catch (error) { console.error(error) }
-	try { LightCounter.init() } catch (error) { console.error(error) }
-	try { WallsCounter.init() } catch (error) { console.error(error) }
-	try { TileCounter.init() } catch (error) { console.error(error) }
-	try { SoundCounter.init() } catch (error) { console.error(error) }
 });
 
 Hooks.once('setup', function () {
@@ -95,4 +94,11 @@ Hooks.once('ready', async function () {
 	try { AltLightNegativeRadius.ready() } catch (error) { console.error(error) }
 	try { LightTemplateManager.ready() } catch (error) { console.error(error) }
 	try { LightingLayerOverride.ready() } catch (error) { console.error(error) }
+	try { ShowLayerControls.ready() } catch (error) { console.error(error) }
+	// WARNING: The *Counters need to go after ShowLayerControls so they don't
+	// accidentally pick up the layer activations during initialization
+	try { LightCounter.ready() } catch (error) { console.error(error) }
+	try { WallsCounter.ready() } catch (error) { console.error(error) }
+	try { TileCounter.ready() } catch (error) { console.error(error) }
+	try { SoundCounter.ready() } catch (error) { console.error(error) }
 });